+++ packages/astro/src/core/compile/compile.ts
@@ -3,9 +3,10 @@ import type { AstroConfig } from '../../@types/astro';
 import type { TransformStyle } from './types';
 
 import { transform } from '@astrojs/compiler';
-import { AstroErrorCodes } from '../errors.js';
+import { AstroErrorCodes } from '../errors/codes.js';
+import { AggregateError, AstroError, CompilerError } from '../errors/errors.js';
 import { prependForwardSlash } from '../path.js';
-import { AggregateError, resolvePath, viteID } from '../util.js';
+import { resolvePath, viteID } from '../util.js';
 import { createStylePreprocessor } from './style.js';
 
 type CompilationCache = Map<string, CompileResult>;
@@ -30,7 +31,7 @@ async function compile({
 	transformStyle,
 }: CompileProps): Promise<CompileResult> {
 	let cssDeps = new Set<string>();
-	let cssTransformErrors: Error[] = [];
+	let cssTransformErrors: AstroError[] = [];
 
 	// Transform from `.astro` to valid `.ts`
 	// use `sourcemap: "both"` so that sourcemap is included in the code
@@ -51,26 +52,51 @@ async function compile({
 			return resolvePath(specifier, filename);
 		},
 	})
-		.catch((err) => {
-			// throw compiler errors here if encountered
-			err.code = err.code || AstroErrorCodes.UnknownCompilerError;
-			throw err;
+		.catch((err: Error) => {
+			// The compiler should be able to handle errors by itself, however
+			// for the rare cases where it can't let's directly throw here with as much info as possible
+			throw new CompilerError({
+				errorCode: AstroErrorCodes.UnknownCompilerError,
+				message: err.message ?? 'Unknown compiler error',
+				stack: err.stack,
+				location: {
+					file: filename,
+				},
+			});
 		})
 		.then((result) => {
+			const compilerError = result.diagnostics.find(
+				// HACK: The compiler currently mistakenly returns the wrong severity for warnings, so we'll also filter by code
+				// https://github.com/withastro/compiler/issues/595
+				(diag) => diag.severity === 1 && diag.code < 2000
+			);
+
+			if (compilerError) {
+				throw new CompilerError({
+					errorCode: compilerError.code,
+					message: compilerError.text,
+					location: {
+						line: compilerError.location.line,
+						column: compilerError.location.column,
+						file: compilerError.location.file,
+					},
+					hint: compilerError.hint ? compilerError.hint : undefined,
+				});
+			}
+
 			switch (cssTransformErrors.length) {
 				case 0:
 					return result;
 				case 1: {
 					let error = cssTransformErrors[0];
-					if (!(error as any).code) {
-						(error as any).code = AstroErrorCodes.UnknownCompilerCSSError;
+					if (!error.errorCode) {
+						error.errorCode = AstroErrorCodes.UnknownCompilerCSSError;
 					}
+
 					throw cssTransformErrors[0];
 				}
 				default: {
-					const aggregateError = new AggregateError(cssTransformErrors);
-					(aggregateError as any).code = AstroErrorCodes.UnknownCompilerCSSError;
-					throw aggregateError;
+					throw new AggregateError({ ...cssTransformErrors[0], errors: cssTransformErrors });
 				}
 			}
 		});