+++ packages/create-astro/src/templates.ts
@@ -1,22 +1,5 @@
 export const TEMPLATES = [
-	{
-		title: 'Just the basics (recommended)',
-		value: 'basics',
-	},
-	{
-		title: 'Blog',
-		value: 'blog',
-	},
-	{
-		title: 'Portfolio',
-		value: 'portfolio',
-	},
-	{
-		title: 'Documentation Site',
-		value: 'docs',
-	},
-	{
-		title: 'Empty project',
-		value: 'minimal',
-	},
+	{ value: 'basics', title: 'a few best practices (recommended)' },
+	{ value: 'blog', title: 'a personal website starter kit' },
+	{ value: 'minimal', title: 'an empty project' },
 ];